requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
h11==0.16.0
idna==3.11
jh2==5.0.10
markdown-it-py==4.0.0
mdurl==0.1.2
Nuitka==2.8.4
ordered-set==4.1.0
Pygments==2.19.2
python-dotenv==1.1.1
qh3==1.5.5
requests==2.32.5
rich==14.2.0
soupsieve==2.8
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
urllib3-future==2.14.905
wassima==2.0.2
zstandard==0.25.0